First pass at moving client logic to api package
