+++ api/pyproject.toml
@@ -18,9 +18,14 @@ classifiers = [
 ]
 dependencies = [
     "backoff==2.2.1",
+    "charset-normalizer>=3.4.1",
     "pandas>=2.0",
     "pydantic>2.0.0",
     "pydantic-settings>2.0.0",
+    "pypdfium2>=4.30.1",
+    "python-docx>=1.1.2",
+    "python-magic>=0.4.27",
+    "python-pptx>=1.0.2",
     "tritonclient",
 ]
