+++ client/pyproject.toml
@@ -18,7 +18,6 @@ classifiers = [
 ]
 dependencies = [
     "build>=1.2.2",
-    "charset-normalizer>=3.4.1",
     "click>=8.1.8",
     "fsspec>=2025.2.0",
     "httpx>=0.28.1",
