+++ client/src/nv_ingest_client/cli/util/processing.py
@@ -19,7 +19,6 @@ from typing import List
 from typing import Tuple
 
 from nv_ingest_client.util.processing import handle_future_result
-from nv_ingest_client.util.util import estimate_page_count
 from PIL import Image
 from tqdm import tqdm
 
@@ -457,7 +456,9 @@ def create_and_process_jobs(
     retry_job_ids: List[str] = []
     job_id_map: Dict[str, str] = {}
     retry_counts: Dict[str, int] = defaultdict(int)
-    file_page_counts: Dict[str, int] = {file: estimate_page_count(file) for file in files}
+    # TODO: Replace this logic ...
+    # file_page_counts: Dict[str, int] = {file: estimate_page_count(file) for file in files}
+    file_page_counts: Dict[str, int] = {}
 
     start_time_ns: int = time.time_ns()
     with tqdm(total=total_files, desc="Processing files", unit="file") as pbar:
